pyproject.toml
[project]
name = "genai-three"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-mistralai>=1.1.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.43.0",
]
